RespiraBakend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0

# CORS
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9

# Production server
gunicorn
dj-database-url
